OrdinaryDiffEqOperatorSplitting

Toolbox to handle split formulations of a wide variety of ODE and DAE problems.

Documentation for OrdinaryDiffEqOperatorSplitting.

OrdinaryDiffEqOperatorSplitting.GenericSplitFunction

OrdinaryDiffEqOperatorSplitting.LieTrotterGodunov

OrdinaryDiffEqOperatorSplitting.OperatorSplittingIntegrator

OrdinaryDiffEqOperatorSplitting.OperatorSplittingProblem

OrdinaryDiffEqOperatorSplitting.step_accept_controller!

OrdinaryDiffEqOperatorSplitting.step_reject_controller!

OrdinaryDiffEqOperatorSplitting.stepsize_controller!

OrdinaryDiffEqOperatorSplitting.OperatorSplittingIntegrator — Type

OperatorSplittingIntegrator <: AbstractODEIntegrator

A variant of ODEIntegrator to perform opeartor splitting.

Derived from https://github.com/CliMA/ClimaTimeSteppers.jl/blob/ef3023747606d2750e674d321413f80638136632/src/integrators.jl.

OrdinaryDiffEqOperatorSplitting.OperatorSplittingProblem — Type

OperatorSplittingProblem(f::AbstractOperatorSplitFunction, u0, tspan, p::Tuple)

OrdinaryDiffEqOperatorSplitting.step_accept_controller! — Method

step_accept_controller!(::OperatorSplittingIntegrator)

Updates _dt of the integrator if the step is accepted and the operator splitting algorithm is adaptive.

OrdinaryDiffEqOperatorSplitting.step_reject_controller! — Method

step_reject_controller!(::OperatorSplittingIntegrator)

Updates _dt of the integrator if the step is rejected and the the operator splitting algorithm is adaptive.

OrdinaryDiffEqOperatorSplitting.stepsize_controller! — Method

stepsize_controller!(::OperatorSplittingIntegrator)

Updates the controller using the current state of the integrator if the operator splitting algorithm is adaptive.