API Reference Overview

OrdinaryDiffEqOperatorSplitting.LieTrotterGodunov — Type

LieTrotterGodunov <: AbstractOperatorSplittingAlgorithm

A first order operator splitting algorithm attributed to [1–3].

OrdinaryDiffEqOperatorSplitting.GenericSplitFunction — Type

GenericSplitFunction(functions::Tuple, solution_indices::Tuple)
GenericSplitFunction(functions::Tuple, solution_indices::Tuple, syncronizers::Tuple)

This type of function describes a set of connected inner functions in mass-matrix form, as usually found in operator splitting procedures.

Automatic sync

We should be able to get rid of the synchronizer and handle the connection of coefficients and solutions in semidiscretize.

Missing docstring.

Missing docstring for OperatorSplittingIntegrator. Check Documenter's build log for details.