API Reference

Problems

OrdinaryDiffEqOperatorSplitting.OperatorSplittingProblem — Type

OperatorSplittingProblem(f::AbstractOperatorSplitFunction, u0, tspan, p::Tuple)

Functions

OrdinaryDiffEqOperatorSplitting.GenericSplitFunction — Type

GenericSplitFunction(functions::Tuple, solution_indices::Tuple)
GenericSplitFunction(functions::Tuple, solution_indices::Tuple, syncronizers::Tuple)

This type of function describes a set of connected inner functions in mass-matrix form, as usually found in operator splitting procedures.

Solver

OrdinaryDiffEqOperatorSplitting.LieTrotterGodunov — Type

LieTrotterGodunov <: AbstractOperatorSplittingAlgorithm

A first order sequential operator splitting algorithm attributed to [1–3].